Add database interface.
Created an interface for database. Makes it more abstract so that implementing different types of database will be easier in the future.
