+++ CatMouseTougePlugin/CatMouseTouge.cs
@@ -4,9 +4,9 @@ using AssettoServer.Server.Configuration;
 using AssettoServer.Server.Plugin;
 using AssettoServer.Shared.Services;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Data.Sqlite;
 using AssettoServer.Network.Tcp;
 using CatMouseTougePlugin.Packets;
+using CatMouseTougePlugin.Database;
 
 namespace CatMouseTougePlugin;
 
@@ -17,8 +17,9 @@ public class CatMouseTouge : CriticalBackgroundService, IAssettoServerAutostart
     private readonly Dictionary<int, EntryCarTougeSession> _instances = [];
     private readonly CSPServerScriptProvider _scriptProvider;
     private readonly CSPClientMessageTypeManager _cspClientMessageTypeManager;
+    private readonly CatMouseTougeConfiguration _configuration;
 
-    public readonly string dbPath = "plugins/CatMouseTougePlugin/database.db";
+    public readonly IDatabase database;
 
     public CatMouseTouge(
         CatMouseTougeConfiguration configuration,
@@ -34,6 +35,7 @@ public class CatMouseTouge : CriticalBackgroundService, IAssettoServerAutostart
         _entryCarTougeSessionFactory = entryCarTougeSessionFactory;
         _scriptProvider = scriptProvider;
         _cspClientMessageTypeManager = cspClientMessageTypeManager;
+        _configuration = configuration;
 
         if (!serverConfiguration.Extra.EnableClientMessages)
         {
@@ -45,7 +47,14 @@ public class CatMouseTouge : CriticalBackgroundService, IAssettoServerAutostart
         ProvideScript("hud.lua");
         ProvideScript("scaling.lua");
 
-        InitializeDatabase(dbPath);
+        // Set up database connection
+        if (_configuration.isDbLocalMode)
+        {
+            // SQLite database.
+            database = new SQLiteDatabase("plugins/CatMouseTougePlugin/database.db");
+        }
+
+        database.InitializeDatabase();
 
         _cspClientMessageTypeManager.RegisterOnlineEvent<EloPacket>(OnEloPacket);
         _cspClientMessageTypeManager.RegisterOnlineEvent<InvitePacket>(OnInvitePacket);
@@ -66,103 +75,11 @@ public class CatMouseTouge : CriticalBackgroundService, IAssettoServerAutostart
 
     internal EntryCarTougeSession GetSession(EntryCar entryCar) => _instances[entryCar.SessionId];
 
-    private void InitializeDatabase(string dbPath)
-    {
-        if (!File.Exists(dbPath))
-        {
-            using var connection = new SqliteConnection($"Data Source={dbPath}");
-            connection.Open();
-            var command = connection.CreateCommand();
-            command.CommandText =
-            @"
-                    CREATE TABLE Players (
-                        PlayerId TEXT PRIMARY KEY,
-                        Rating INTEGER,
-                        RacesCompleted INTEGER
-                    );
-            ";
-            command.ExecuteNonQuery();
-        }
-    }
-
     private void OnClientConnected(ACTcpClient client, EventArgs args)
     {
         // Check if the player is registered in the database
         string playerId = client.Guid.ToString();
-
-        // Query the database with clientId.
-        using var connection = new SqliteConnection($"Data Source={dbPath}");
-        connection.Open();
-
-        // First, check if the player exists
-        var checkCommand = connection.CreateCommand();
-        checkCommand.CommandText = "SELECT COUNT(*) FROM Players WHERE PlayerId = @PlayerId";
-        checkCommand.Parameters.AddWithValue("@PlayerId", playerId);
-
-        int playerExists = Convert.ToInt32(checkCommand.ExecuteScalar());
-
-        // If player doesn't exist, add them with default values
-        if (playerExists == 0)
-        {
-            var insertCommand = connection.CreateCommand();
-            insertCommand.CommandText = "INSERT INTO Players (PlayerId, Rating, RacesCompleted) VALUES (@PlayerId, @Rating, @RacesCompleted)";
-            insertCommand.Parameters.AddWithValue("@PlayerId", playerId);
-            insertCommand.Parameters.AddWithValue("@Rating", 1000); // Default ELO rating
-            insertCommand.Parameters.AddWithValue("@RacesCompleted", 0);
-            insertCommand.ExecuteNonQuery();
-        }
-    }
-
-    public int GetPlayerElo(string playerId)
-    {
-        using var connection = new SqliteConnection($"Data Source={dbPath}");
-        connection.Open();
-
-        using var command = connection.CreateCommand();
-        command.CommandText = @"
-        SELECT Rating
-        FROM Players
-        WHERE PlayerId = $playerId
-    ";
-        command.Parameters.AddWithValue("$playerId", playerId);
-
-        var result = command.ExecuteScalar();
-
-        if (result != null && int.TryParse(result.ToString(), out int rating))
-        {
-            return rating;
-        }
-        else
-        {
-            throw new Exception($"Player with ID {playerId} not found in the database.");
-        }
-    }
-
-    public (int Rating, int RacesCompleted) GetPlayerStats(string playerId)
-    {
-        using var connection = new SqliteConnection($"Data Source={dbPath}");
-        connection.Open();
-
-        using var command = connection.CreateCommand();
-        command.CommandText = @"
-        SELECT Rating, RacesCompleted
-        FROM Players
-        WHERE PlayerId = $playerId;
-    ";
-        command.Parameters.AddWithValue("$playerId", playerId);
-
-        using var reader = command.ExecuteReader();
-
-        if (reader.Read())
-        {
-            int rating = reader.GetInt32(0);           // Rating
-            int racesCompleted = reader.GetInt32(1);   // RacesCompleted
-            return (rating, racesCompleted);
-        }
-        else
-        {
-            throw new Exception($"Player with ID {playerId} not found in the database.");
-        }
+        database.CheckPlayer(playerId);
     }
 
     private void ProvideScript(string scriptName)
@@ -175,11 +92,11 @@ public class CatMouseTouge : CriticalBackgroundService, IAssettoServerAutostart
         _scriptProvider.AddScript(reconnectScript, scriptName);
     }
 
-    private void OnEloPacket(ACTcpClient client, EloPacket packet)
+    private async void OnEloPacket(ACTcpClient client, EloPacket packet)
     {
         int elo = 1000; // Default Elo if player not found
         string playerId = client.Guid.ToString();
-        elo = GetPlayerElo(playerId);
+        elo = await database.GetPlayerElo(playerId);
 
         client.SendPacket(new EloPacket { Elo = elo });
     }
@@ -312,6 +229,5 @@ public class CatMouseTouge : CriticalBackgroundService, IAssettoServerAutostart
     {
         client?.SendPacket(new NotificationPacket { Message = message });
     }
-
 }
